+++ R/global_sessions_config.R
@@ -20,7 +20,6 @@
 #' Make sure to set \code{admin_mode = FALSE} before deploying your app.
 #' @param is_invite_required \code{TRUE} by default.  Whether or not to require the user to have an
 #' invite before registering/signing in
-#' @param api_url the API url.  Defaults to \code{"https://api.polished.tech"}.
 #' @param sign_in_providers the sign in providers to enable.  Valid values are \code{"google"}
 #' \code{"email"}, \code{"microsoft"}, and/or \code{"facebook"}. Defaults to \code{"email"}.
 #' @param is_email_verification_required \code{TRUE} by default.  Whether or not to require the user to
@@ -54,7 +53,6 @@ global_sessions_config <- function(
   firebase_config = NULL,
   admin_mode = FALSE,
   is_invite_required = TRUE,
-  api_url = "https://api.polished.tech",
   sign_in_providers = "email",
   is_email_verification_required = TRUE,
   is_auth_required = TRUE,
@@ -65,22 +63,17 @@ global_sessions_config <- function(
     stop("invalid `api_key` argument passed to `global_sessions_config()`", call. = FALSE)
   }
 
-  if (!(length(api_url) == 1 && is.character(api_url))) {
-    stop("invalid `api_url` argument passed to `global_sessions_config()`", call. = FALSE)
-  }
-
-
+  current_polished_options <- getOption("polished")
   # get the app uid
   res <- httr::GET(
-    url = paste0(api_url, "/apps"),
+    url = paste0(current_polished_options$api_url, "/apps"),
     query = list(
       app_name = app_name
     ),
     httr::authenticate(
       user = api_key,
       password = ""
-    ),
-    config = list(http_version = 0)
+    )
   )
 
   app <- jsonlite::fromJSON(
@@ -91,7 +84,9 @@ global_sessions_config <- function(
     stop(app, call. = FALSE)
   }
 
-  if (length(app) == 0) {
+  app <- tibble::as_tibble(app)
+
+  if (identical(nrow(app), 0L)) {
     stop(paste0("app_name `", app_name, "` does not exist"), call. = FALSE)
   }
 
@@ -104,14 +99,16 @@ global_sessions_config <- function(
     stop("invalid `sentry_dsn` argument passed to `global_sessions_config()`", call. = FALSE)
   }
 
-  options("polished" = list(
-    api_key = api_key,
-    api_url = api_url,
-    app_uid = app$uid,
-    app_name = app_name,
-    app_name_display = app_name_display,
-    sentry_dsn = sentry_dsn
-  ))
+
+
+
+  options_out <- current_polished_options
+  options_out$api_key <- api_key
+  options_out$app_uid <- app$uid
+  options_out$app_name_display <- app_name_display
+  options_out$sentry_dsn <- sentry_dsn
+  options("polished" = options_out)
+
 
   .global_sessions$config(
     firebase_config = firebase_config,