+++ R/package_deps.R
@@ -25,8 +25,7 @@
 #' dir <- system.file("examples", "polished_example_01", package = "polished")
 #' pkg_deps <- polished:::get_package_deps(dir)
 #'
-#' @importFrom automagic get_dependent_packages get_package_details
-#' @importFrom cli cli_alert_warning cli_alert_danger cat_bullet
+#' @importFrom automagic get_package_details
 #' @importFrom dplyr %>%
 #' @importFrom purrr safely map_depth pluck compact map
 get_package_deps <- function(
@@ -40,12 +39,12 @@ get_package_deps <- function(
   }
 
   # detect R package dependencies
-  init_pkg_names <- automagic::get_dependent_packages(app_dir)
+  init_pkg_names <- get_dependent_packages(app_dir)
 
   # return if no detections
   if (length(init_pkg_names) == 0) {
-    cli::cli_alert_warning("warning: no packages found in specified directory")
-    return(invisible(NULL))
+    warning("no packages found in specified directory", call. = FALSE)
+    invisible(NULL)
   }
 
 
@@ -69,7 +68,8 @@ get_package_deps <- function(
   hold <- hold[!(names(hold) %in% errors)]
 
   if (length(errors) > 0 && verbose) {
-    cli::cli_alert_danger("Silently removing detected invalid packages: {errors}")
+    removed_packages <- paste(errors, collapse = ", ")
+    warning(paste0("Silently removing detected invalid packages: ", removed_packages), call. = FALSE)
   }
 
   purrr::map_depth(hold, 1, purrr::pluck, "result") %>%
@@ -77,4 +77,40 @@ get_package_deps <- function(
       if (length(x) == 0) return(NULL) else return(x)
     }) %>%
     purrr::compact()
+}
+
+#' get packages required to run R code
+#'
+#' Note: this function is copied from the \code{automagic} R package.  We are including it in
+#' \code{polished} while we await the merging of this PR \url{https://github.com/cole-brokamp/automagic/pull/17}
+#' and a new CRAN release of \code{automagic}.
+#'
+#' @details parses all R and Rmd files in a directory and uses \code{automagic::parse_packages}
+#'     to find all R packages required for the code to run
+#'
+#' @param directory folder to search for R and Rmd files
+#'
+#' @return a vector of package names
+#'
+#' @importFrom automagic parse_packages
+#'
+get_dependent_packages <- function(directory = getwd()) {
+
+  fls <- list.files(
+    path = directory,
+    pattern = '^.*\\.R$|^.*\\.Rmd$',
+    full.names = TRUE,
+    recursive = TRUE,
+    ignore.case = TRUE
+  )
+
+  pkg_names <- unlist(sapply(fls, automagic::parse_packages))
+  pkg_names <- unique(pkg_names)
+
+  if (length(pkg_names)==0) {
+    message('warning: no packages found in specified directory')
+    return(invisible(NULL))
+  }
+
+  return(unname(pkg_names))
 }