+++ tests/testthat.R
@@ -1,0 +1,6 @@
+library(testthat)
+library(polished)
+library(httr)
+
+test_check("polished")
+#testthat::test_local()